admincu/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
bcrypt==3.1.7
cached-property==1.4.0
cairocffi==0.8.0
CairoSVG==2.1.3
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
cryptography==3.2.1
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.2.19
django-afip==7.1.2
django-anymail==2.2
django-appconf==1.0.2
django-crontab==0.7.1
django-dbbackup==3.2.0
django-filter==2.0.0
django-formtools==2.1
django-import-export==1.0.0
django-mailer==1.2.4
django-maintenance-mode==0.9.0
django-mercadopago==5.1.1
django-postman==3.6.1
django-render-block==0.5
django-renderpdf==3.0.0
django-templated-email==2.2.0
django-termsandconditions==2.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
future==0.16.0
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
IPy==0.83
isodate==0.6.0
jdcal==1.4
lockfile==0.12.2
mercadopago==0.3.5
odfpy==1.3.6
openpyxl==2.5.3
paramiko==2.7.2
pdfrw==0.4
pycparser==2.18
PyNaCl==1.4.0
pyOpenSSL==17.5.0
Pyphen==0.9.4
python-barcode==0.8.1
python-dateutil==2.7.1
pytz==2018.3
PyYAML==3.12
qrcode==6.1
requests==2.18.4
requests-toolbelt==0.8.0
setuptools-git==1.2
setuptools-scm==2.0.0
six==1.11.0
sqlparse==0.4.1
supertools==1.0.1
tablib==0.12.1
tinycss2==0.6.1
unicodecsv==0.14.1
urllib3==1.22
WeasyPrint==0.42.2
webencodings==0.5.1
xlrd==1.1.0
xlwt==1.3.0
zeep==2.5.0
mysqlclient==1.3.13
